% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-documentation.R
\name{BirthDeathRates_df}
\alias{BirthDeathRates_df}
\title{Changes in Human Birth and Death Rates in India Over the 20th Century}
\format{
A data frame with 27 observations and 3 variables:
\describe{
  \item{Year}{Year of observation (factor)}
  \item{Birth.rate}{Birth rate (numeric)}
  \item{death.rate}{Death rate (numeric)}
}
}
\source{
Data taken from the \pkg{gpk} package version 1.0
}
\usage{
data(BirthDeathRates_df)
}
\description{
This dataset, BirthDeathRates_df, is a data frame containing data on human
birth and death rates in India over the 20th century. It includes the year,
birth rate, and death rate for each recorded period.
}
\details{
The dataset name has been kept as 'BirthDeathRates_df' to avoid confusion with other datasets
in the R ecosystem. This naming convention helps distinguish this dataset as part of the
IndiAPIs package and assists users in identifying its specific characteristics.
The suffix 'df' indicates that the dataset is a data frame object. The original content has not been modified
in any way.
}
