% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-documentation.R
\name{BombayPlague1905_df}
\alias{BombayPlague1905_df}
\title{Weekly deaths from bubonic plague in Bombay in 1905-06}
\format{
A data frame with 32 observations and 2 variables:
\describe{
  \item{Week}{Week number of the observation period (integer)}
  \item{CumulativeDeaths}{Cumulative number of plague deaths (integer)}
}
}
\source{
Data taken from the \pkg{primer} package version 1.2.0
}
\usage{
data(BombayPlague1905_df)
}
\description{
This dataset, BombayPlague1905_df, is a data frame containing the number of plague deaths per week
in Bombay in 1905–06. The data was originally reported by Kermack and McCormick (1927). Bombay is the former
name for the Indian coastal city Mumbai, which is the capital of Maharashtra and one of the largest cities
in the world.
}
\details{
The dataset name has been kept as 'BombayPlague1905_df' to avoid confusion with other datasets
in the R ecosystem. This naming convention helps distinguish this dataset as part of the
IndiAPIs package and assists users in identifying its specific characteristics.
The suffix 'df' indicates that the dataset is a data frame object. The original content has not been modified
in any way.
}
