% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-documentation.R
\name{BurdwanRiceYield_df}
\alias{BurdwanRiceYield_df}
\title{Yearly Rice Yield Data in Burdwan District, West Bengal}
\format{
A data frame with 39 observations and 2 variables:
\describe{
  \item{Year}{Year of observation (character)}
  \item{burdwan}{Rice yield in tonnes per hectare (numeric)}
}
}
\source{
Data taken from the \pkg{weatherindices} package version 0.1.0
}
\usage{
data(BurdwanRiceYield_df)
}
\description{
This dataset, BurdwanRiceYield_df, is a data frame containing yearly rice yield
data for the Burdwan district of West Bengal, India, over a period of 39 years.
It includes the year and the yield in tonnes per hectare for each recorded year.
}
\details{
The dataset name has been kept as 'BurdwanRiceYield_df' to avoid confusion with other datasets
in the R ecosystem. This naming convention helps distinguish this dataset as part of the
IndiAPIs package and assists users in identifying its specific characteristics.
The suffix 'df' indicates that the dataset is a data frame object. The original content has not been modified
in any way.
}
