% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-documentation.R
\name{DelhiPotatoPrices_ts}
\alias{DelhiPotatoPrices_ts}
\title{Monthly Average Potato Price of Delhi Market (India)}
\format{
A time series with 127 time points and 1 variable:
\describe{
  \item{Delhi}{Monthly average potato price in the Delhi market (numeric)}
}
}
\source{
Data taken from the \pkg{stlELM} package version 0.1.1
}
\usage{
data(DelhiPotatoPrices_ts)
}
\description{
This dataset, DelhiPotatoPrices_ts, is a time series containing the monthly average potato prices
of the Delhi market from January 2010 to July 2020.
}
\details{
The dataset name has been kept as 'DelhiPotatoPrices_ts' to avoid confusion with other datasets
in the R ecosystem. This naming convention helps distinguish this dataset as part of the
IndiAPIs package and assists users in identifying its specific characteristics.
The suffix 'ts' indicates that the dataset is a time series object. The original content has not been modified
in any way.
}
