% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-documentation.R
\name{indianPopulation_tbl_df}
\alias{indianPopulation_tbl_df}
\title{Indian Population (Census and Projections) by States}
\format{
A tibble with 36 observations and 8 variables:
\describe{
  \item{code}{Numeric state code (numeric)}
  \item{abbr}{State abbreviation (character)}
  \item{state}{Full state name (character)}
  \item{pop_1901}{Population in the year 1901 (numeric)}
  \item{pop_1951}{Population in the year 1951 (numeric)}
  \item{pop_2011}{Population in the year 2011 (numeric)}
  \item{pop_2023}{Population in the year 2023 (numeric)}
  \item{pop_2024}{Population in the year 2024 (numeric)}
}
}
\source{
Data taken from the \pkg{mapindia} package version 1.0.1
}
\usage{
data(indianPopulation_tbl_df)
}
\description{
This dataset, indianPopulation_tbl_df, is a tibble containing census data and
population projections for Indian states across multiple years. It includes
state codes, abbreviations, names, and population figures for the years 1901,
1951, 2011, 2023, and 2024.
}
\details{
The dataset name has been kept as 'indianPopulation_tbl_df' to avoid confusion with other datasets
in the R ecosystem. This naming convention helps distinguish this dataset as part of the
IndiAPIs package and assists users in identifying its specific characteristics.
The suffix 'tbl_df' indicates that the dataset is a tibble object. The original content has not been modified
in any way.
}
