% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/indiapis-package.R
\docType{package}
\name{IndiAPIs}
\alias{IndiAPIs}
\alias{IndiAPIs-package}
\title{IndiAPIs: Access Indian Data via Public APIs and Curated Datasets}
\description{
{
This package provides functions to access data from public RESTful APIs including 'World Bank API', and 'REST Countries API', retrieving real-time or historical data related to India, such as economic indicators, and international demographic and geopolitical indicators. Additionally, the package includes one of the largest curated collections of open datasets focused on India, covering topics such as population, economy, weather, politics, health, biodiversity, sports, agriculture, cybercrime, infrastructure, and more.
}
}
\details{
IndiAPIs: Access Indian Data via Public APIs and Curated Datasets

\if{html}{\figure{logo.png}{options: style='float: right' alt='logo' width='120'}}

Access Indian Data via Public APIs and Curated Datasets.
}
\seealso{
{
Useful links:
\itemize{
  \item \url{https://github.com/lightbluetitan/indiapis}
}
}
}
\author{
{
\strong{Maintainer}: Renzo Caceres Rossi \email{arenzocaceresrossi@gmail.com}
}
}
