% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-documentation.R
\name{IndiaPopulation_dt}
\alias{IndiaPopulation_dt}
\title{List of places, abbreviations, and populations in India}
\format{
A data.table with 39 observations and 3 variables:
\describe{
  \item{place}{Name of the state or union territory (character)}
  \item{abbrev}{Abbreviation for the state or union territory (character)}
  \item{population}{Population in 2019 projection (numeric)}
}
}
\source{
Data taken from the \pkg{covid19india} package version 0.1.4
}
\usage{
data(IndiaPopulation_dt)
}
\description{
This dataset, IndiaPopulation_dt, is a data table containing the names of states and union territories
in India along with their respective abbreviations and populations. The dataset also includes the
total population of India. These are 2019 projections as reported in the Unique Identification
Authority of India 2019-2020 Annual Report.
}
\details{
The dataset name has been kept as 'IndiaPopulation_dt' to avoid confusion with other datasets
in the R ecosystem. This naming convention helps distinguish this dataset as part of the
IndiAPIs package and assists users in identifying its specific characteristics.
The suffix 'dt' indicates that the dataset is a data.table object. The original content has not been modified
in any way.
}
